refactor: split now-playing.command.ts into co-located handler modules
Extracts the ~30 interaction handlers from the 2056-line
now-playing.command.ts into six co-located @Discord handler modules
under src/commands/now-playing/ (add, platform, sort, remove, edit-note,
nav). The command class now retains only the three @Slash entry points
and their slash-only helpers (showSingle, showEveryone,
resolveNowPlayingGameByTitle, replaceNowPlayingMessageInCurrentChannel),
dropping from 2056 to 606 lines. All custom IDs and restart-resume
identifiers are preserved unchanged. Pre-existing dead private method
promptEditNowPlayingNote is retained as-is (out of scope to remove).

Closes #925
