+++ src/commands/now-playing.command.ts
@@ -3,12 +3,6 @@ import {
   type CommandInteraction,
   type User,
   MessageFlags,
-  ModalSubmitInteraction,
-  StringSelectMenuBuilder,
-  StringSelectMenuInteraction,
-  ButtonInteraction,
-  type ActionRow,
-  type MessageActionRowComponent,
   type Message,
 } from "discord.js";
 import {
@@ -16,37 +10,24 @@ import {
   Slash,
   SlashOption,
   SlashGroup,
-  SelectMenuComponent,
-  ButtonComponent,
-  ModalComponent,
 } from "discordx";
-import Member, { type IMemberNowPlayingEntry } from "../classes/Member.js";
+import Member from "../classes/Member.js";
 import {
   extractErrorMessage,
-  getModalField,
   safeDeferReply,
-  safeDeferUpdate,
   safeReply,
   safeUpdate,
-  safeUserFetch,
   sanitizeUserInput,
   type AnyRepliable,
-  replyIfNotOwner,
 } from "../functions/InteractionUtils.js";
 import type { IGame } from "../types/GameTypes.js";
-import Game from "../classes/Game.js";
 import {
-  buildActionButton,
-  buildButtonRow,
   buildUserHeaderContainer,
-  buildSelectRow,
 } from "../functions/uiComponents.js";
 import {
   buildComponentsV2Flags,
-  buildComponentsV2EditFlags,
   buildTextContainer,
   buildTextReply,
-  buildTitledContainer,
 } from "../functions/ComponentsV2Utils.js";
 import {
   autocompleteGameCompletionPlatform,
@@ -54,85 +35,40 @@ import {
   resolveGameCompletionPlatformId,
 } from "./game-completion/completion-autocomplete.utils.js";
 import { parseTitleWithYear } from "../functions/GameTitleAutocompleteUtils.js";
-import { STANDARD_PLATFORM_IDS } from "../config/standardPlatforms.js";
-import {
-  NOW_PLAYING_HELP_TEXTS,
-} from "./now-playing-help.js";
 
 import { renderUsernameWithEmoji } from "../services/UserEmojiService.js";
-import { isPositiveInt } from "../utilities/ValidationUtils.js";
-import {
-  DISCORD_SELECT_OPTIONS_MAX,
-  truncateLabel,
-} from "../config/textLimits.js";
-import { assertCustomIdSegments, parseCustomIdSegmentsMin } from "../utilities/CustomIdUtils.js";
 import { safeIgnore } from "../utilities/AsyncUtils.js";
 
 import {
   MAX_NOW_PLAYING_NOTE_LEN,
   NOW_PLAYING_SEARCH_LIMIT,
-  NOW_PLAYING_NOTE_INPUT_ID,
   NOW_PLAYING_NOTE_MODAL_MAX_FIELDS,
-  NOW_PLAYING_ADD_MODAL_ID,
-  NOW_PLAYING_ADD_TITLE_INPUT_ID,
-  NOW_PLAYING_ADD_NOTE_INPUT_ID,
-  NOW_PLAYING_GALLERY_MAX,
   NOW_PLAYING_LIST_EDIT_PREFIX,
 } from "./now-playing/nowPlayingIds.js";
 import {
-  type NowPlayingAddSession,
-} from "./now-playing/nowPlayingTypes.js";
-import {
-  nowPlayingAddSessions,
-  nowPlayingAddPlatformSessions,
   nowPlayingListContexts,
-  nowPlayingOwnerMenu,
-  clearNowPlayingAddSession,
   trackNowPlayingListContext,
-  setNowPlayingListContext,
   NOW_PLAYING_CONTEXT_TTL_MS,
 } from "./now-playing/nowPlayingContexts.js";
 import {
-  buildNowPlayingSortStateToken,
-  parseNowPlayingSortStateToken,
-  encodeNowPlayingSortState,
-  parseNowPlayingPlatformStateToken,
-  encodeNowPlayingPlatformState,
-  buildNowPlayingPlatformStateFromCurrent,
-  formatEntryTitleWithPlatform,
   getDisplayNowPlayingEntries,
 } from "../functions/NowPlayingUtils.js";
 import {
   buildNowPlayingListContainer,
   buildNowPlayingMessageContainer,
-  buildComponentPayload,
-  buildNowPlayingAttachments,
   buildNowPlayingListPayload,
-  buildNowPlayingManageRow,
-  returnToNowPlayingEditMenu,
-  buildNowPlayingEditInitialComponents,
   withPmNowPlayingList,
   refreshNowPlayingListFromContext,
   trimTextDisplayContent,
   buildNowPlayingMemberSelect,
-  buildNowPlayingRemoveComponents,
-  buildNowPlayingEditPlatformComponents,
-  buildNowPlayingSortComponents,
 } from "./now-playing/nowPlayingListRenderer.js";
 import {
-  buildEditNoteModal,
   buildEditNotesModal,
-  buildNowPlayingAddModal,
 } from "./now-playing/nowPlayingModals.js";
 import {
   deleteEligibleNowPlayingMessageInCurrentChannel,
 } from "./now-playing/nowPlayingMessageService.js";
-import { promptNowPlayingAddPlatformSelection } from "./now-playing/nowPlayingAddService.js";
 import GamePlatformRegionService from "../classes/GamePlatformRegionService.js";
-import {
-  startNowPlayingIgdbImport,
-  startNowPlayingIgdbImportFromInteraction,
-} from "./now-playing/nowPlayingIgdbImport.service.js";
 import GameSearchService from "../classes/GameSearchService.js";
 
 @Discord()
@@ -391,126 +327,6 @@ export class NowPlayingCommand {
     });
   }
 
-  @ModalComponent({ id: NOW_PLAYING_ADD_MODAL_ID })
-  async handleAddNowPlayingModal(interaction: ModalSubmitInteraction): Promise<void> {
-    const query = getModalField(interaction, NOW_PLAYING_ADD_TITLE_INPUT_ID);
-    const noteRaw = getModalField(interaction, NOW_PLAYING_ADD_NOTE_INPUT_ID);
-    if (!query) {
-      const container = buildTextContainer("Please provide a title to search.");
-      await safeReply(interaction, {
-        components: [container],
-        flags: buildComponentsV2Flags(true),
-      });
-      return;
-    }
-    if (noteRaw.length > MAX_NOW_PLAYING_NOTE_LEN) {
-      const container = buildTextContainer(`Note must be ${MAX_NOW_PLAYING_NOTE_LEN} characters or fewer.`);
-      await safeReply(interaction, {
-        components: [container],
-        flags: buildComponentsV2Flags(true),
-      });
-      return;
-    }
-
-    try {
-      const results = await GameSearchService.searchGames(query);
-      if (!results.length) {
-        await startNowPlayingIgdbImportFromInteraction(
-          interaction,
-          {
-            userId: interaction.user.id,
-            query,
-            note: noteRaw.length ? noteRaw : null,
-          },
-          "reply",
-        );
-        return;
-      }
-      const sessionId = `np-${interaction.user.id}`;
-      const session: NowPlayingAddSession = {
-        userId: interaction.user.id,
-        query,
-        note: noteRaw.length ? noteRaw : null,
-      };
-      nowPlayingAddSessions.set(sessionId, session);
-
-      const options: Array<{ label: string; value: string; description?: string }> =
-        results.slice(0, 23).map((g) => ({
-        label: g.title.substring(0, 100),
-        value: String(g.id),
-      }));
-
-      options.push({
-        label: "Import another game from IGDB",
-        value: "import-igdb",
-        description: "Search IGDB and import a new GameDB entry",
-      });
-
-      const selectId = `nowplaying-add-select:${sessionId}`;
-      const selectRow = buildSelectRow(
-        new StringSelectMenuBuilder()
-          .setCustomId(selectId)
-          .setPlaceholder("Select the game to add")
-          .addOptions(options),
-      );
-
-      const contentLines = [
-        "## Now Playing Add",
-        "Select a game to add to your Now Playing list:",
-      ];
-      if (results.length > options.length - 1) {
-        contentLines.push(`Showing first ${options.length - 1} results.`);
-      }
-      const content = trimTextDisplayContent(contentLines.join("\n"));
-      const container = buildTextContainer(content)
-        .addActionRowComponents(selectRow.toJSON());
-
-      const reply = await safeReply(interaction, {
-        components: [container],
-        flags: buildComponentsV2Flags(true),
-        withResponse: true,
-      } as any);
-      const replyMessage = reply?.resource?.message ?? null;
-
-      session.timeoutId = setTimeout(async () => {
-        try {
-          if (!nowPlayingAddSessions.has(sessionId)) {
-            return;
-          }
-          if (!replyMessage) {
-            return;
-          }
-          const hasMatchingSelect = replyMessage.components.some(
-            (row: ActionRow<MessageActionRowComponent>) => {
-            if (!("components" in row)) return false;
-            const actionRow = row as ActionRow<MessageActionRowComponent>;
-            return actionRow.components.some(
-              (component) =>
-                "customId" in component && component.customId === selectId,
-            );
-          });
-          if (!hasMatchingSelect) return;
-
-          const timeoutContainer = buildTextContainer("Timed out waiting for a selection. No changes made.");
-          await safeReply(interaction, {
-            components: [timeoutContainer],
-            flags: buildComponentsV2Flags(true),
-          });
-          clearNowPlayingAddSession(sessionId);
-        } catch {
-          // ignore
-        }
-      }, 60_000);
-    } catch (err: any) {
-      const msg = extractErrorMessage(err);
-      const container = buildTextContainer(`Could not add to Now Playing: ${msg}`);
-      await safeReply(interaction, {
-        components: [container],
-        flags: buildComponentsV2Flags(true),
-      });
-    }
-  }
-
   private async resolveNowPlayingGameByTitle(searchTerm: string): Promise<IGame | null> {
     const parsed = parseTitleWithYear(searchTerm);
     const normalizedSearchTerm = parsed.title.trim();
@@ -545,265 +361,6 @@ export class NowPlayingCommand {
     return null;
   }
 
-  @SelectMenuComponent({ id: /^nowplaying-add-select:.+$/ })
-  async handleAddNowPlayingSelect(interaction: StringSelectMenuInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 1);
-    if (!segs) return;
-    const [sessionId] = segs;
-    const session = nowPlayingAddSessions.get(sessionId);
-    const ownerId = session?.userId;
-
-    if (!session || interaction.user.id !== ownerId) {
-      const container = buildTextContainer("This add prompt isn't for you.");
-      await safeReply(interaction, {
-        components: [container],
-        flags: buildComponentsV2Flags(true),
-      });
-      return;
-    }
-
-    const choice = interaction.values[0];
-    if (choice === "import-igdb") {
-      await startNowPlayingIgdbImport(interaction, session);
-      return;
-    }
-    const gameId = Number(choice);
-    if (!isPositiveInt(gameId)) {
-      const container = buildTextContainer("Invalid selection. Please try again.");
-      await safeUpdate(interaction, {
-        components: [container],
-        flags: buildComponentsV2Flags(true),
-      });
-      clearNowPlayingAddSession(sessionId);
-      return;
-    }
-
-    try {
-      await promptNowPlayingAddPlatformSelection(
-        interaction,
-        sessionId,
-        ownerId,
-        gameId,
-        session.note,
-        "update",
-      );
-    } catch (err: any) {
-      const msg = extractErrorMessage(err);
-      const container = buildTextContainer(`Could not add to Now Playing: ${msg}`);
-      await safeUpdate(interaction, {
-        components: [container],
-        flags: buildComponentsV2Flags(true),
-      });
-      clearNowPlayingAddSession(sessionId);
-    }
-  }
-
-  @SelectMenuComponent({ id: /^nowplaying-add-platform-select:[^:]+$/ })
-  async handleAddNowPlayingPlatformSelect(interaction: StringSelectMenuInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 1);
-    if (!segs) return;
-    const [platformSessionId] = segs;
-    const session = nowPlayingAddPlatformSessions.get(platformSessionId);
-    if (!session) {
-      const container = buildTextContainer("This platform prompt has expired.");
-      await safeReply(interaction, {
-        components: [container],
-        flags: buildComponentsV2Flags(true),
-      });
-      return;
-    }
-
-    if (await replyIfNotOwner(interaction, session.userId, "This platform prompt isn't for you.")) return;
-
-    const platformId = Number(interaction.values?.[0]);
-    if (!isPositiveInt(platformId)) {
-      const container = buildTextContainer("Invalid platform selection.");
-      await safeReply(interaction, {
-        components: [container],
-        flags: buildComponentsV2Flags(true),
-      });
-      return;
-    }
-
-    await safeDeferUpdate(interaction);
-    const loadingContainer = buildTextContainer("## Now Loading\nGenerating cover layout and loading the selected member list...");
-    await safeUpdate(interaction, {
-      components: [loadingContainer],
-      flags: buildComponentsV2Flags(true),
-    });
-
-    try {
-      await Member.addNowPlaying(session.userId, session.gameId, platformId, session.note);
-      const trimmedSessionNote = session.note?.trim();
-      if (trimmedSessionNote) {
-        await Member.addGameJournalEntry({
-          userId: session.userId,
-          gameId: session.gameId,
-          body: trimmedSessionNote,
-        });
-      }
-      nowPlayingAddPlatformSessions.delete(platformSessionId);
-      clearNowPlayingAddSession(session.sourceSessionId);
-      const list = await Member.getNowPlaying(session.userId);
-      const payload = await buildNowPlayingListPayload(
-        interaction.user,
-        list,
-        interaction.guildId,
-      );
-      const refreshed = await refreshNowPlayingListFromContext(interaction, session.userId);
-      if (refreshed) {
-        return;
-      } else {
-        await safeUpdate(interaction, {
-          components: payload.components,
-          files: payload.files,
-          flags: buildComponentsV2Flags(true),
-        });
-      }
-    } catch (err: any) {
-      const msg = extractErrorMessage(err);
-      const container = buildTextContainer(`Could not add to Now Playing: ${msg}`);
-      await safeUpdate(interaction, {
-        components: [container],
-        flags: buildComponentsV2Flags(true),
-      });
-      nowPlayingAddPlatformSessions.delete(platformSessionId);
-      clearNowPlayingAddSession(session.sourceSessionId);
-    }
-  }
-
-  private async promptRemoveNowPlaying(
-    interaction: AnyRepliable,
-    mode: "reply" | "update" = "reply",
-  ): Promise<void> {
-    if (mode === "reply") {
-      await safeDeferReply(interaction, { flags: buildComponentsV2Flags(true) });
-    }
-    const userId = interaction.user.id;
-    const useDeferredEditPath = mode === "update" &&
-      Boolean((interaction as any).__rpgDeferred ?? (interaction as any).deferred);
-    const isEphemeral = mode === "update"
-      ? ((interaction as any).message?.flags?.has(MessageFlags.Ephemeral) ?? false)
-      : true;
-    try {
-      const entries = getDisplayNowPlayingEntries(await Member.getNowPlaying(userId));
-      if (!entries.length) {
-        const container = buildTextContainer("Your Now Playing list is empty.");
-        const pmComponents = await withPmNowPlayingList(
-          userId,
-          interaction.guildId,
-          [container],
-        );
-        if (mode === "update" && !useDeferredEditPath) {
-          await safeUpdate(interaction, {
-            components: pmComponents,
-            flags: buildComponentsV2Flags(isEphemeral),
-          });
-        } else if (mode === "update") {
-          await safeReply(interaction, {
-            components: pmComponents,
-            flags: buildComponentsV2Flags(true),
-          });
-        } else {
-          await safeReply(interaction, {
-            components: pmComponents,
-            flags: buildComponentsV2Flags(true),
-          });
-        }
-        return;
-      }
-
-      const includeImages = interaction.guildId != null;
-      const { files, thumbnailsByGameId } = await buildNowPlayingAttachments(
-        entries,
-        NOW_PLAYING_GALLERY_MAX,
-        includeImages,
-      );
-      const components = buildNowPlayingRemoveComponents(
-        entries,
-        userId,
-        thumbnailsByGameId,
-      );
-      const pmComponents = await withPmNowPlayingList(
-        userId,
-        interaction.guildId,
-        components,
-      );
-      if (mode === "update" && !useDeferredEditPath) {
-        await safeUpdate(interaction, {
-          ...buildComponentPayload(pmComponents as any, files),
-          flags: buildComponentsV2Flags(isEphemeral),
-        });
-      } else if (mode === "update") {
-        await safeReply(interaction, {
-          ...buildComponentPayload(pmComponents as any, files),
-          flags: buildComponentsV2Flags(true),
-        });
-      } else {
-        await safeReply(interaction, {
-          ...buildComponentPayload(pmComponents as any, files),
-          flags: buildComponentsV2Flags(true),
-        });
-      }
-    } catch (err: any) {
-      const msg = extractErrorMessage(err);
-      const container = buildTextContainer(`Could not remove from Now Playing: ${msg}`);
-      const pmComponents = await withPmNowPlayingList(
-        userId,
-        interaction.guildId,
-        [container],
-      );
-      if (mode === "update" && !useDeferredEditPath) {
-        await safeUpdate(interaction, {
-          components: pmComponents,
-          flags: buildComponentsV2Flags(isEphemeral),
-        });
-      } else if (mode === "update") {
-        await safeReply(interaction, {
-          components: pmComponents,
-          flags: buildComponentsV2Flags(true),
-        });
-      } else {
-        await safeReply(interaction, {
-          components: pmComponents,
-          flags: buildComponentsV2Flags(true),
-        });
-      }
-    }
-  }
-
-  private async promptSortNowPlayingButtons(
-    interaction: ButtonInteraction,
-    ownerId: string,
-  ): Promise<void> {
-    const isEphemeral = interaction.message.flags?.has(MessageFlags.Ephemeral) ?? false;
-    const entries = getDisplayNowPlayingEntries(
-      await Member.getNowPlaying(ownerId),
-    ).slice(0, 10);
-    if (!entries.length) {
-      const container = buildTextContainer("Your Now Playing list is empty.");
-      const pmComponents = await withPmNowPlayingList(
-        ownerId, interaction.guildId, [container],
-      );
-      await safeUpdate(interaction, {
-        components: pmComponents,
-        flags: buildComponentsV2Flags(isEphemeral),
-      });
-      return;
-    }
-    const stateToken = buildNowPlayingSortStateToken(entries.length);
-    const components = buildNowPlayingSortComponents(entries, ownerId, stateToken);
-    const pmComponents = await withPmNowPlayingList(
-      ownerId,
-      interaction.guildId,
-      components,
-    );
-    await safeUpdate(interaction, {
-      components: pmComponents,
-      flags: buildComponentsV2Flags(isEphemeral),
-    });
-  }
-
   private async promptEditNowPlayingNote(
     interaction: AnyRepliable,
     mode: "reply" | "update" = "reply",
@@ -821,7 +378,7 @@ export class NowPlayingCommand {
         [container],
       );
       if (mode === "update" && "update" in interaction) {
-        await safeUpdate(interaction, { 
+        await safeUpdate(interaction, {
           components: pmComponents, flags: buildComponentsV2Flags(true) });
       } else {
         await safeReply(interaction, {
@@ -846,7 +403,7 @@ export class NowPlayingCommand {
         [container],
       );
       if (mode === "update" && "update" in interaction) {
-        await safeUpdate(interaction, { 
+        await safeUpdate(interaction, {
           components: pmComponents, flags: buildComponentsV2Flags(true) });
       } else {
         await safeReply(interaction, {
@@ -867,879 +424,6 @@ export class NowPlayingCommand {
       await safeReply(interaction, buildTextReply(`Discord modals support up to ${NOW_PLAYING_NOTE_MODAL_MAX_FIELDS} note fields at once. ` +
           "I opened the first set. Submit, then use Edit Notes again for the rest.", true));
     }
-  }
-
-  private async promptEditNowPlayingPlatform(
-    interaction: AnyRepliable,
-    mode: "reply" | "update" = "reply",
-  ): Promise<void> {
-    if (mode === "reply") {
-      await safeDeferReply(interaction, { flags: MessageFlags.Ephemeral });
-    }
-
-    const entries = getDisplayNowPlayingEntries(await Member.getNowPlaying(interaction.user.id));
-    if (!entries.length) {
-      const container = buildTextContainer("Your Now Playing list is empty.");
-      const pmComponents = await withPmNowPlayingList(
-        interaction.user.id,
-        interaction.guildId,
-        [container],
-      );
-      if (mode === "update" && "update" in interaction) {
-        await safeUpdate(interaction, { 
-          components: pmComponents, flags: buildComponentsV2Flags(true) });
-      } else {
-        await safeReply(interaction, {
-          components: pmComponents,
-          flags: MessageFlags.Ephemeral,
-        });
-      }
-      return;
-    }
-
-    const platformOptions = await this.getNowPlayingEditPlatformOptions(entries);
-    const stateToken = buildNowPlayingPlatformStateFromCurrent(entries, platformOptions);
-    const components = buildNowPlayingEditPlatformComponents(
-      entries,
-      interaction.user.id,
-      platformOptions,
-      stateToken,
-    );
-    const pmComponents = await withPmNowPlayingList(
-      interaction.user.id,
-      interaction.guildId,
-      components,
-    );
-
-    if (mode === "update" && "update" in interaction) {
-      await safeUpdate(interaction, {
-        components: pmComponents,
-        flags: buildComponentsV2Flags(true),
-      });
-      return;
-    }
-    await safeReply(interaction, {
-      components: pmComponents,
-      flags: buildComponentsV2Flags(true),
-    });
-  }
-
-  private async getNowPlayingEditPlatformOptions(
-    entries: IMemberNowPlayingEntry[],
-  ): Promise<Array<Array<{ label: string; value: string; platformId: number }>>> {
-    const limitedEntries = entries.slice(0, 10);
-    const optionsPerEntry = await Promise.all(
-      limitedEntries.map(async (entry) => {
-        const platforms = await GamePlatformRegionService.getPlatformsForGameWithStandard(
-          entry.gameId,
-          STANDARD_PLATFORM_IDS,
-        );
-        const uniqueById = new Map<number, { id: number; name: string }>();
-        platforms.forEach((platform) => {
-          if (!uniqueById.has(platform.id)) {
-            uniqueById.set(platform.id, platform);
-          }
-        });
-        const deduped = Array.from(uniqueById.values()).slice(0, DISCORD_SELECT_OPTIONS_MAX);
-        if (!deduped.length && entry.platformId) {
-          deduped.push({
-            id: entry.platformId,
-            name: entry.platformName ?? "Current Platform",
-          });
-        }
-        return deduped.map((platform, optionIndex) => ({
-          label: truncateLabel(platform.name),
-          value: String(optionIndex),
-          platformId: platform.id,
-        }));
-      }),
-    );
-    return optionsPerEntry;
-  }
-
-  private async promptNowPlayingEditPlatformSelection(
-    interaction: AnyRepliable,
-    ownerId: string,
-    gameId: number,
-    mode: "reply" | "update" = "reply",
-  ): Promise<void> {
-    const game = await Game.getGameById(gameId);
-    if (!game) {
-      const container = buildTextContainer("That game could not be found.");
-      if (mode === "update" && "update" in interaction) {
-        await safeUpdate(interaction, { 
-          components: [container], flags: buildComponentsV2Flags(true) });
-      } else {
-        await safeReply(interaction, {
-          components: [container],
-          flags: buildComponentsV2Flags(true),
-        });
-      }
-      return;
-    }
-
-    const platforms = await GamePlatformRegionService
-      .getPlatformsForGameWithStandard(gameId, STANDARD_PLATFORM_IDS);
-    if (!platforms.length) {
-      const container = buildTextContainer("No platform data is available for this game.");
-      if (mode === "update" && "update" in interaction) {
-        await safeUpdate(interaction, { 
-          components: [container], flags: buildComponentsV2Flags(true) });
-      } else {
-        await safeReply(interaction, {
-          components: [container],
-          flags: buildComponentsV2Flags(true),
-        });
-      }
-      return;
-    }
-
-    const options = platforms.slice(0, DISCORD_SELECT_OPTIONS_MAX).map((platform) => ({
-      label: truncateLabel(platform.name),
-      value: String(platform.id),
-    }));
-    const select = new StringSelectMenuBuilder()
-      .setCustomId(`nowplaying-edit-platform-select:${ownerId}:${gameId}`)
-      .setPlaceholder("Select the platform")
-      .addOptions(options);
-    const content = platforms.length > options.length
-      ? `Select the platform for **${game.title}** (showing first ${options.length}).`
-      : `Select the platform for **${game.title}**.`;
-    const container = buildTextContainer(content);
-    const payload = {
-      components: [
-        container,
-        buildSelectRow(select),
-      ],
-      flags: buildComponentsV2Flags(true),
-    };
-    const pmComponents = await withPmNowPlayingList(
-      ownerId,
-      interaction.guildId,
-      payload.components,
-    );
-    if (mode === "update" && "update" in interaction) {
-      await safeUpdate(interaction, { ...payload, components: pmComponents });
-    } else {
-      await safeReply(interaction, { ...payload, components: pmComponents });
-    }
-  }
-
-  @SelectMenuComponent({ id: /^nowplaying-edit-platform-select:\d+:\d+$/ })
-  async handleNowPlayingEditPlatformSelect(
-    interaction: StringSelectMenuInteraction,
-  ): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 2);
-    if (!segs) return;
-    const [ownerId, gameIdRaw] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This platform prompt isn't for you.")) return;
-
-    const gameId = Number(gameIdRaw);
-    const platformId = Number(interaction.values?.[0]);
-    if (!isPositiveInt(gameId) || !isPositiveInt(platformId)) {
-      await safeReply(interaction, buildTextReply("Invalid platform selection.", true));
-      return;
-    }
-
-    const updated = await Member.updateNowPlayingPlatform(ownerId, gameId, platformId);
-    if (!updated) {
-      await safeReply(interaction, buildTextReply("Could not update that platform.", true));
-      return;
-    }
-
-    safeIgnore(refreshNowPlayingListFromContext(interaction, ownerId));
-    await returnToNowPlayingEditMenu(interaction, ownerId);
-  }
-
-  @SelectMenuComponent({ id: /^nowplaying-edit-platform-slot:\d+:\d+:[a-z0-9_]+$/ })
-  async handleEditPlatformSlot(interaction: StringSelectMenuInteraction): Promise<void> {
-    const isEphemeral = interaction.message.flags?.has(MessageFlags.Ephemeral) ?? false;
-    const segs = assertCustomIdSegments(interaction, 3);
-    if (!segs) return;
-    const [ownerId, slotRaw, stateToken] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This platform prompt isn't for you.")) return;
-
-    const slotIndex = Number(slotRaw);
-    const selectedOptionIndex = Number(interaction.values?.[0]);
-    if (
-      !Number.isInteger(slotIndex) ||
-      slotIndex < 0 ||
-      !Number.isInteger(selectedOptionIndex) ||
-      selectedOptionIndex < 0
-    ) {
-      await safeReply(interaction, buildTextReply("Invalid selection.", true));
-      return;
-    }
-
-    const entries = getDisplayNowPlayingEntries(await Member.getNowPlaying(ownerId)).slice(0, 10);
-    const platformOptions = await this.getNowPlayingEditPlatformOptions(entries);
-    const parsed = parseNowPlayingPlatformStateToken(stateToken, entries.length);
-    if (!parsed || slotIndex >= entries.length || selectedOptionIndex > 24) {
-      await safeReply(interaction, buildTextReply("This platform form has expired. Open Edit Platform again.", true));
-      return;
-    }
-    if (selectedOptionIndex >= (platformOptions[slotIndex]?.length ?? 0)) {
-      await safeReply(interaction, buildTextReply("Invalid platform selection for that game.", true));
-      return;
-    }
-
-    parsed[slotIndex] = selectedOptionIndex;
-    const components = buildNowPlayingEditPlatformComponents(
-      entries,
-      ownerId,
-      platformOptions,
-      encodeNowPlayingPlatformState(parsed),
-    );
-    const pmComponents = await withPmNowPlayingList(ownerId, interaction.guildId, components);
-    await safeUpdate(interaction, {
-      components: pmComponents,
-      flags: buildComponentsV2Flags(isEphemeral),
-    });
-  }
-
-  @SelectMenuComponent({ id: /^nowplaying-edit-note-select:\d+$/ })
-  async handleEditNoteSelect(interaction: StringSelectMenuInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 1);
-    if (!segs) return;
-    const [ownerId] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This note prompt isn't for you.")) return;
-
-    const gameId = Number(interaction.values?.[0]);
-    if (!isPositiveInt(gameId)) {
-      await safeReply(interaction, buildTextReply("Invalid selection.", true));
-      return;
-    }
-
-    const currentEntries = await Member.getNowPlayingEntries(ownerId);
-    const currentEntry = currentEntries.find((entry) => entry.gameId === gameId);
-    if (!currentEntry) {
-      await safeReply(interaction, buildTextReply("Entry not found.", true));
-      return;
-    }
-
-    await interaction.showModal(
-      buildEditNoteModal(
-        ownerId,
-        gameId,
-        formatEntryTitleWithPlatform(currentEntry),
-        currentEntry.note ?? null,
-      ),
-    ).catch(async () => {
-      await safeReply(interaction, buildTextReply("Unable to open the note form right now.", true));
-    });
-  }
-
-  @ButtonComponent({ id: /^nowplaying-edit-note-direct:\d+:\d+$/ })
-  async handleEditNoteDirect(interaction: ButtonInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 2);
-    if (!segs) return;
-    const [ownerId, gameIdRaw] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This note prompt isn't for you.")) return;
-    const gameId = Number(gameIdRaw);
-    if (!isPositiveInt(gameId)) {
-      await safeReply(interaction, buildTextReply("Invalid selection.", true));
-      return;
-    }
-    const currentEntries = await Member.getNowPlayingEntries(ownerId);
-    const currentEntry = currentEntries.find((entry) => entry.gameId === gameId);
-    if (!currentEntry) {
-      await safeReply(interaction, buildTextReply("Entry not found.", true));
-      return;
-    }
-    if (currentEntry.journalEnabled) {
-      await safeReply(interaction, buildTextReply("This game uses Game Journal for notes. Use the Journal button to add entries.", true));
-      return;
-    }
-    setNowPlayingListContext(ownerId, interaction.message);
-    await interaction.showModal(
-      buildEditNoteModal(
-        ownerId,
-        gameId,
-        formatEntryTitleWithPlatform(currentEntry),
-        currentEntry.note ?? null,
-      ),
-    ).catch(async () => {
-      await safeReply(interaction, buildTextReply("Unable to open the note form right now.", true));
-    });
-  }
-
-  @SelectMenuComponent({ id: /^nowplaying-sort-slot:\d+:\d+:[a-z0-9_]+$/ })
-  async handleNowPlayingSortSlot(interaction: StringSelectMenuInteraction): Promise<void> {
-    const isEphemeral = interaction.message.flags?.has(MessageFlags.Ephemeral) ?? false;
-    const segs = assertCustomIdSegments(interaction, 3);
-    if (!segs) return;
-    const [ownerId, slotRaw, stateToken] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This sort prompt isn't for you.")) return;
-
-    const slotIndex = Number(slotRaw);
-    if (!Number.isInteger(slotIndex) || slotIndex < 0) {
-      const container = buildTextContainer("Invalid selection.");
-      await safeReply(interaction, {
-        components: [container],
-        flags: buildComponentsV2Flags(true),
-      });
-      return;
-    }
-
-    try {
-      const entries = getDisplayNowPlayingEntries(await Member.getNowPlaying(ownerId)).slice(0, 10);
-      const parsed = parseNowPlayingSortStateToken(stateToken, entries.length);
-      const selectedValue = interaction.values[0] ?? "";
-      const selectedIndex = Number(selectedValue);
-      if (
-        !parsed ||
-        !Number.isInteger(selectedIndex) ||
-        selectedIndex < 0 ||
-        selectedIndex >= entries.length ||
-        slotIndex >= entries.length
-      ) {
-        const container = buildTextContainer("This sort form has expired. Open Sort again.");
-        await safeUpdate(interaction, {
-          components: [container],
-          flags: buildComponentsV2Flags(isEphemeral),
-        });
-        return;
-      }
-
-      for (let i = 0; i < parsed.length; i += 1) {
-        if (i !== slotIndex && parsed[i] === selectedIndex) {
-          parsed[i] = -1;
-        }
-      }
-      parsed[slotIndex] = selectedIndex;
-      const components = buildNowPlayingSortComponents(
-        entries,
-        ownerId,
-        encodeNowPlayingSortState(parsed),
-      );
-      const pmComponents = await withPmNowPlayingList(
-        ownerId, interaction.guildId, components,
-      );
-      await safeUpdate(interaction, {
-        components: pmComponents,
-        flags: buildComponentsV2Flags(isEphemeral),
-      });
-    } catch {
-      const container = buildTextContainer("Could not update the sort form right now.");
-      safeIgnore(safeUpdate(interaction, {
-        components: [container],
-        flags: buildComponentsV2Flags(isEphemeral),
-      }));
-    }
-  }
-
-  @ButtonComponent({ id: /^nowplaying-sort-save:\d+:[a-z0-9_]+$/ })
-  async handleNowPlayingSortSave(interaction: ButtonInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 2);
-    if (!segs) return;
-    const [ownerId, stateToken] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This sort prompt isn't for you.")) return;
-    await safeDeferUpdate(interaction);
-    const isEphemeral = interaction.message.flags?.has(MessageFlags.Ephemeral) ?? false;
-    const responseFlags = buildComponentsV2Flags(isEphemeral);
-
-    const entries = getDisplayNowPlayingEntries(await Member.getNowPlaying(ownerId)).slice(0, 10);
-    const parsed = parseNowPlayingSortStateToken(stateToken, entries.length);
-    if (!parsed) {
-      const container = buildTextContainer("This sort form has expired. Open Sort again.");
-      const pmComponents = await withPmNowPlayingList(
-        ownerId, interaction.guildId, [container],
-      );
-      await safeReply(interaction, { components: pmComponents, flags: responseFlags });
-      return;
-    }
-    if (parsed.some((value) => value < 0)) {
-      const components = buildNowPlayingSortComponents(
-        entries,
-        ownerId,
-        stateToken,
-        "Assign a title to every visible position before saving.",
-      );
-      const pmComponents = await withPmNowPlayingList(
-        ownerId, interaction.guildId, components,
-      );
-      await safeReply(interaction, { components: pmComponents, flags: responseFlags });
-      return;
-    }
-    if (new Set(parsed).size !== parsed.length) {
-      const components = buildNowPlayingSortComponents(
-        entries,
-        ownerId,
-        stateToken,
-        "Each title can only be used once. Remove duplicate assignments and try again.",
-      );
-      const pmComponents = await withPmNowPlayingList(
-        ownerId, interaction.guildId, components,
-      );
-      await safeReply(interaction, { components: pmComponents, flags: responseFlags });
-      return;
-    }
-
-    const loadingContainer = buildTextContainer("## Now Loading\nSaving sort order and generating cover layout...");
-    await safeUpdate(interaction, { components: [loadingContainer], flags: responseFlags });
-
-    const orderedIds = parsed.map((index) => entries[index].gameId);
-    const updated = await Member.updateNowPlayingSort(ownerId, orderedIds);
-    if (!updated) {
-      const container = buildTextContainer("Could not update the sort order.");
-      const pmComponents = await withPmNowPlayingList(
-        ownerId, interaction.guildId, [container],
-      );
-      await safeReply(interaction, { components: pmComponents, flags: responseFlags });
-      return;
-    }
-
-    safeIgnore(refreshNowPlayingListFromContext(interaction, ownerId));
-    await returnToNowPlayingEditMenu(interaction, ownerId);
-  }
-
-  @ButtonComponent({ id: /^nowplaying-sort-reset:\d+$/ })
-  async handleNowPlayingSortReset(interaction: ButtonInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 1);
-    if (!segs) return;
-    const [ownerId] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This sort prompt isn't for you.")) return;
-    await safeDeferUpdate(interaction);
-    const isEphemeral = interaction.message.flags?.has(MessageFlags.Ephemeral) ?? false;
-    const responseFlags = buildComponentsV2Flags(isEphemeral);
-    const entries = getDisplayNowPlayingEntries(await Member.getNowPlaying(ownerId)).slice(0, 10);
-    const stateToken = buildNowPlayingSortStateToken(entries.length);
-    const components = buildNowPlayingSortComponents(entries, ownerId, stateToken);
-    const pmComponents = await withPmNowPlayingList(ownerId, interaction.guildId, components);
-    await safeReply(interaction, { components: pmComponents, flags: responseFlags });
-  }
-
-  @ModalComponent({ id: /^nowplaying-note-modal:\d+(?::\d+)?$/ })
-  async handleEditNoteModal(interaction: ModalSubmitInteraction): Promise<void> {
-    await safeDeferReply(interaction, { flags: MessageFlags.Ephemeral });
-    const segs = parseCustomIdSegmentsMin(interaction.customId, 1);
-    if (!segs) return;
-    const [ownerId, legacyGameIdRaw = null] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This note prompt isn't for you.")) return;
-
-    let updated = false;
-    if (legacyGameIdRaw) {
-      const gameId = Number(legacyGameIdRaw);
-      if (!isPositiveInt(gameId)) {
-        await safeReply(interaction, buildTextReply("Invalid selection.", true));
-        return;
-      }
-
-      const noteInput = getModalField(interaction, NOW_PLAYING_NOTE_INPUT_ID);
-      const note = noteInput.trim();
-      const nextNote = note ? note : null;
-      if (note && note.length > MAX_NOW_PLAYING_NOTE_LEN) {
-        await safeReply(interaction, buildTextReply(`Note must be ${MAX_NOW_PLAYING_NOTE_LEN} characters or fewer.`, true));
-        return;
-      }
-
-      updated = await Member.updateNowPlayingNote(ownerId, gameId, nextNote);
-    } else {
-      const currentEntries = await Member.getNowPlayingEntries(ownerId);
-      const updateCandidates = currentEntries.slice(0, NOW_PLAYING_NOTE_MODAL_MAX_FIELDS);
-
-      for (const entry of updateCandidates) {
-        if (!entry.gameId || entry.journalEnabled) {
-          continue;
-        }
-        const fieldId = `${NOW_PLAYING_NOTE_INPUT_ID}:${entry.gameId}`;
-        let noteInput = "";
-        try {
-          noteInput = getModalField(interaction, fieldId);
-        } catch {
-          noteInput = "";
-        }
-        const note = noteInput.trim();
-        if (note.length > MAX_NOW_PLAYING_NOTE_LEN) {
-          await safeReply(interaction, buildTextReply(`Note must be ${MAX_NOW_PLAYING_NOTE_LEN} characters or fewer.`, true));
-          return;
-        }
-        const nextNote = note ? note : null;
-        const changed = await Member.updateNowPlayingNote(ownerId, entry.gameId, nextNote);
-        updated = changed || updated;
-      }
-    }
-    if (updated) {
-      const refreshed = await refreshNowPlayingListFromContext(interaction, ownerId);
-      if (!interaction.guildId && interaction.message) {
-        try {
-          const dmComponents = await buildNowPlayingEditInitialComponents(ownerId);
-          await interaction.message.edit({
-            components: dmComponents,
-            flags: buildComponentsV2Flags(false),
-          });
-          safeIgnore(interaction.deleteReply());
-          return;
-        } catch {
-          // Fall through to existing fallback response if DM message edit fails.
-        }
-      }
-      if (refreshed) {
-        safeIgnore(interaction.deleteReply());
-        return;
-      }
-      const list = await Member.getNowPlaying(ownerId);
-      const payload = await buildNowPlayingListPayload(
-        interaction.user,
-        list,
-        interaction.guildId,
-      );
-      await safeReply(interaction, {
-        components: payload.components,
-        files: payload.files,
-        flags: buildComponentsV2Flags(true),
-      });
-      return;
-    }
-    await safeReply(interaction, buildTextReply("Could not update that entry.", true));
-  }
-
-  @SelectMenuComponent({ id: /^nowplaying-delete-note-select:\d+$/ })
-  async handleDeleteNoteSelect(interaction: StringSelectMenuInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 1);
-    if (!segs) return;
-    const [ownerId] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This note prompt isn't for you.")) return;
-
-    const gameId = Number(interaction.values?.[0]);
-    if (!isPositiveInt(gameId)) {
-      await safeReply(interaction, buildTextReply("Invalid selection.", true));
-      return;
-    }
-
-    const currentEntries = await Member.getNowPlayingEntries(ownerId);
-    const currentEntry = currentEntries.find((entry) => entry.gameId === gameId);
-    const currentNote = currentEntry?.note ? currentEntry.note : "No note set.";
-    if (!currentEntry) {
-      await safeReply(interaction, buildTextReply("Entry not found.", true));
-      return;
-    }
-
-    const noteBody = currentEntry.note ? `> ${currentNote}` : "No note set.";
-    const container = buildTitledContainer(`Delete Note: ${currentEntry.title}`, noteBody);
-
-    const row = buildButtonRow(
-      buildActionButton("delete", `nowplaying-delete-note-confirm:${ownerId}:${gameId}:yes`, "Delete Note"),
-      buildActionButton("cancel", `nowplaying-delete-note-confirm:${ownerId}:${gameId}:no`),
-    );
-
-    await safeUpdate(interaction, {
-      components: [container, row],
-      flags: buildComponentsV2EditFlags(),
-    });
-  }
-
-  @ButtonComponent({ id: /^nowplaying-delete-note-confirm:\d+:\d+:(yes|no)$/ })
-  async handleDeleteNoteConfirm(interaction: ButtonInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 3);
-    if (!segs) return;
-    const [ownerId, gameIdRaw, choice] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This note prompt isn't for you.")) return;
-
-    if (choice === "no") {
-      safeIgnore(safeUpdate(interaction, {
-        content: "Cancelled.",
-        components: [],
-      }));
-      return;
-    }
-
-    const gameId = Number(gameIdRaw);
-    if (!isPositiveInt(gameId)) {
-      await safeReply(interaction, buildTextReply("Invalid selection.", true));
-      return;
-    }
-
-    const updated = await Member.updateNowPlayingNote(ownerId, gameId, null);
-    safeIgnore(safeUpdate(interaction, {
-      content: updated ? "Note deleted." : "Could not update that entry.",
-      components: [],
-    }));
-  }
-
-  @ButtonComponent({ id: /^np-remove:[^:]+:\d+$/ })
-  async handleRemoveNowPlayingButton(interaction: ButtonInteraction): Promise<void> {
-    const isEphemeral = interaction.message.flags?.has(MessageFlags.Ephemeral) ?? false;
-    const segs = assertCustomIdSegments(interaction, 2);
-    if (!segs) return;
-    const [ownerId, gameIdRaw] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This remove prompt isn't for you.")) return;
-
-    const gameId = Number(gameIdRaw);
-    if (!isPositiveInt(gameId)) {
-      const container = buildTextContainer("Invalid selection.");
-      await safeReply(interaction, {
-        components: [container],
-        flags: buildComponentsV2Flags(true),
-      });
-      return;
-    }
-
-    try {
-      const removed = await Member.removeNowPlaying(ownerId, gameId);
-      if (!removed) {
-        const container = buildTextContainer("Failed to remove that game (it may have been removed already).");
-        await safeReply(interaction, {
-          components: [container],
-          flags: buildComponentsV2Flags(true),
-        });
-        return;
-      }
-      safeIgnore(refreshNowPlayingListFromContext(interaction, ownerId));
-
-      const entries = getDisplayNowPlayingEntries(await Member.getNowPlaying(ownerId));
-      if (!entries.length) {
-        const container = buildTextContainer("Your Now Playing list is empty.");
-        await safeUpdate(interaction, {
-          components: [container],
-          flags: buildComponentsV2Flags(isEphemeral),
-        });
-        return;
-      }
-      const includeImages = interaction.guildId != null;
-      const { files, thumbnailsByGameId } = await buildNowPlayingAttachments(
-        entries,
-        NOW_PLAYING_GALLERY_MAX,
-        includeImages,
-      );
-      const components = buildNowPlayingRemoveComponents(
-        entries,
-        ownerId,
-        thumbnailsByGameId,
-      );
-      const pmComponents = await withPmNowPlayingList(
-        ownerId,
-        interaction.guildId,
-        components,
-      );
-      await safeUpdate(interaction, {
-        ...buildComponentPayload(pmComponents as any, files),
-        flags: buildComponentsV2Flags(isEphemeral),
-      });
-    } catch (err: any) {
-      const msg = extractErrorMessage(err);
-      const container = buildTextContainer(`Could not remove from Now Playing: ${msg}`);
-      await safeReply(interaction, {
-        components: [container],
-        flags: buildComponentsV2Flags(true),
-      });
-    }
-  }
-
-  @ButtonComponent({ id: /^nowplaying-list-edit:\d+$/ })
-  async handleNowPlayingListEdit(interaction: ButtonInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 1);
-    if (!segs) return;
-    const [ownerId] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "Only the owner of this Now Playing list can use Edit.")) return;
-
-    setNowPlayingListContext(ownerId, interaction.message);
-    await nowPlayingOwnerMenu.show(
-      interaction,
-      ownerId,
-      [await buildNowPlayingManageRow(ownerId)],
-    );
-  }
-
-  @ButtonComponent({ id: /^nowplaying-help:[a-z-]+:\d+$/ })
-  async handleNowPlayingHelp(interaction: ButtonInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 2);
-    if (!segs) return;
-    const [screenType, ownerId] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This help button isn't for you.")) return;
-    const helpText = NOW_PLAYING_HELP_TEXTS[screenType]
-      ?? "No help available for this screen.";
-    const container = buildTextContainer(helpText);
-    await safeReply(interaction, {
-      components: [container],
-      flags: buildComponentsV2Flags(true),
-    });
-  }
-
-  @ButtonComponent({ id: /^nowplaying-edit-menu-sort:\d+$/ })
-  async handleNowPlayingEditMenuSort(interaction: ButtonInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 1);
-    if (!segs) return;
-    const [ownerId] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This edit menu isn't for you.")) return;
-    await this.promptSortNowPlayingButtons(interaction, ownerId);
-  }
-
-  @ButtonComponent({ id: /^nowplaying-edit-menu-platform:\d+$/ })
-  async handleNowPlayingEditMenuPlatform(interaction: ButtonInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 1);
-    if (!segs) return;
-    const [ownerId] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This edit menu isn't for you.")) return;
-    await this.promptEditNowPlayingPlatform(interaction, "update");
-  }
-
-  @ButtonComponent({ id: /^nowplaying-edit-menu-remove:\d+$/ })
-  async handleNowPlayingEditMenuRemove(interaction: ButtonInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 1);
-    if (!segs) return;
-    const [ownerId] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This edit menu isn't for you.")) return;
-    await this.promptRemoveNowPlaying(interaction, "update");
-  }
-
-  @ButtonComponent({ id: /^nowplaying-list-add:\d+$/ })
-  async handleNowPlayingListAdd(interaction: ButtonInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 1);
-    if (!segs) return;
-    const [ownerId] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This add prompt isn't for you.")) return;
-    setNowPlayingListContext(ownerId, interaction.message);
-    safeIgnore(interaction.showModal(buildNowPlayingAddModal()));
-  }
-
-  @ButtonComponent({ id: /^nowplaying-list-edit-platform:\d+$/ })
-  async handleNowPlayingListEditPlatform(interaction: ButtonInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 1);
-    if (!segs) return;
-    const [ownerId] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This platform prompt isn't for you.")) return;
-    setNowPlayingListContext(ownerId, interaction.message);
-    await this.promptEditNowPlayingPlatform(interaction, "update");
-  }
-
-  @ButtonComponent({ id: /^np-edit-platform:\d+:\d+$/ })
-  async handleNowPlayingEditPlatformPick(interaction: ButtonInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 2);
-    if (!segs) return;
-    const [ownerId, gameIdRaw] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This platform prompt isn't for you.")) return;
-    const gameId = Number(gameIdRaw);
-    if (!isPositiveInt(gameId)) {
-      await safeReply(interaction, buildTextReply("Invalid selection.", true));
-      return;
-    }
-    await this.promptNowPlayingEditPlatformSelection(interaction, ownerId, gameId, "update");
-  }
-
-  @ButtonComponent({ id: /^nowplaying-edit-platform-save:\d+:[a-z0-9_]+$/ })
-  async handleNowPlayingEditPlatformSave(interaction: ButtonInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 2);
-    if (!segs) return;
-    const [ownerId, stateToken] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This platform prompt isn't for you.")) return;
-
-    await safeDeferUpdate(interaction);
-    const isEphemeral = interaction.message.flags?.has(MessageFlags.Ephemeral) ?? false;
-    const responseFlags = buildComponentsV2Flags(isEphemeral);
-    const entries = getDisplayNowPlayingEntries(await Member.getNowPlaying(ownerId)).slice(0, 10);
-    const platformOptions = await this.getNowPlayingEditPlatformOptions(entries);
-    const parsed = parseNowPlayingPlatformStateToken(stateToken, entries.length);
-    if (!parsed) {
-      await safeReply(
-        interaction,
-        buildTextReply("This platform form has expired. Open Edit Platform again.", isEphemeral),
-      );
-      return;
-    }
-    if (parsed.some((value) => value < 0)) {
-      const components = buildNowPlayingEditPlatformComponents(
-        entries,
-        ownerId,
-        platformOptions,
-        stateToken,
-        "Assign a platform for every visible game before saving.",
-      );
-      const pmComponents = await withPmNowPlayingList(
-        ownerId, interaction.guildId, components,
-      );
-      await safeReply(interaction, { components: pmComponents, flags: responseFlags });
-      return;
-    }
-
-    for (let slotIndex = 0; slotIndex < entries.length; slotIndex += 1) {
-      const selectedOptionIndex = parsed[slotIndex];
-      const option = platformOptions[slotIndex]?.[selectedOptionIndex];
-      const gameId = entries[slotIndex]?.gameId;
-      if (!option || !gameId) {
-        await safeReply(
-          interaction,
-          buildTextReply(
-            "One or more selected platforms are invalid. Please review and try again.",
-            isEphemeral,
-          ),
-        );
-        return;
-      }
-      const updated = await Member.updateNowPlayingPlatform(ownerId, gameId, option.platformId);
-      if (!updated) {
-        await safeReply(
-          interaction,
-          buildTextReply(
-            `Could not update platform for ${entries[slotIndex].title}.`,
-            isEphemeral,
-          ),
-        );
-        return;
-      }
-    }
-    safeIgnore(refreshNowPlayingListFromContext(interaction, ownerId));
-    await returnToNowPlayingEditMenu(interaction, ownerId);
-  }
-
-  @ButtonComponent({ id: /^nowplaying-edit-platform-reset:\d+$/ })
-  async handleNowPlayingEditPlatformReset(interaction: ButtonInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 1);
-    if (!segs) return;
-    const [ownerId] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This platform prompt isn't for you.")) return;
-    await safeDeferUpdate(interaction);
-    const isEphemeral = interaction.message.flags?.has(MessageFlags.Ephemeral) ?? false;
-    const responseFlags = buildComponentsV2Flags(isEphemeral);
-    const entries = getDisplayNowPlayingEntries(await Member.getNowPlaying(ownerId)).slice(0, 10);
-    const platformOptions = await this.getNowPlayingEditPlatformOptions(entries);
-    const stateTokenReset = buildNowPlayingPlatformStateFromCurrent(entries, platformOptions);
-    const components = buildNowPlayingEditPlatformComponents(
-      entries,
-      ownerId,
-      platformOptions,
-      stateTokenReset,
-    );
-    const pmComponents = await withPmNowPlayingList(ownerId, interaction.guildId, components);
-    await safeReply(interaction, { components: pmComponents, flags: responseFlags });
-  }
-
-  @ButtonComponent({ id: /^nowplaying-list-sort:\d+$/ })
-  async handleNowPlayingListSort(interaction: ButtonInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 1);
-    if (!segs) return;
-    const [ownerId] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This sort prompt isn't for you.")) return;
-    setNowPlayingListContext(ownerId, interaction.message);
-    await this.promptSortNowPlayingButtons(interaction, ownerId);
-  }
-
-  @ButtonComponent({ id: /^nowplaying-list-remove:\d+$/ })
-  async handleNowPlayingListRemove(interaction: ButtonInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 1);
-    if (!segs) return;
-    const [ownerId] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This remove prompt isn't for you.")) return;
-    setNowPlayingListContext(ownerId, interaction.message);
-    await this.promptRemoveNowPlaying(interaction, "update");
-  }
-
-  @ButtonComponent({ id: /^nowplaying-remove-done:\d+$/ })
-  async handleNowPlayingRemoveDone(interaction: ButtonInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 1);
-    if (!segs) return;
-    const [ownerId] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This remove prompt isn't for you.")) return;
-    await returnToNowPlayingEditMenu(interaction, ownerId);
-  }
-
-  @ButtonComponent({ id: /^nowplaying-list-cancel:\d+$/ })
-  async handleNowPlayingListCancel(interaction: ButtonInteraction): Promise<void> {
-    const segs = assertCustomIdSegments(interaction, 1);
-    if (!segs) return;
-    const [ownerId] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This prompt isn't for you.")) return;
-    await returnToNowPlayingEditMenu(interaction, ownerId);
   }
 
   async showSingle(
@@ -1828,63 +512,6 @@ export class NowPlayingCommand {
     }
   }
 
-  @SelectMenuComponent({ id: /^nowplaying-all-select(?::v1)?$/ })
-  async handleNowPlayingAllSelect(interaction: StringSelectMenuInteraction): Promise<void> {
-    const selectedUserId = interaction.values?.[0];
-    if (!selectedUserId) return;
-    const isEphemeral = interaction.message.flags?.has(MessageFlags.Ephemeral) ?? false;
-
-    const loadingContainer = buildTextContainer("## Now Loading\nGenerating cover layout and loading the selected member list...");
-    await safeUpdate(interaction, {
-      components: [loadingContainer],
-      flags: buildComponentsV2Flags(isEphemeral),
-    });
-
-    const entries = await Member.getNowPlaying(selectedUserId);
-    const target =
-      (await safeUserFetch(interaction.client, selectedUserId)) ??
-      interaction.user;
-
-    if (!entries.length) {
-      const ownerName = target.displayName ?? target.username ?? target.id;
-      const header = buildUserHeaderContainer(
-        selectedUserId,
-        ownerName,
-        "Now Playing",
-        `${NOW_PLAYING_LIST_EDIT_PREFIX}:${selectedUserId}`,
-      );
-      const container = buildNowPlayingMessageContainer(
-        "Now Playing - Everyone",
-        `No Now Playing entries found for ${renderUsernameWithEmoji(selectedUserId, ownerName)}.`,
-      );
-      const updated = await safeReply(interaction, {
-        components: [header, container],
-      });
-      trackNowPlayingListContext(updated as Message<boolean>, {
-        view: "everyone-selected",
-        selectedUserId,
-      });
-      return;
-    }
-
-    const sortedEntries = getDisplayNowPlayingEntries(entries);
-    const payload = await buildNowPlayingListPayload(
-      target,
-      sortedEntries,
-      interaction.guildId,
-      false,
-      true,
-    );
-    const updated = await safeReply(interaction, {
-      components: payload.components,
-      files: payload.files,
-    });
-    trackNowPlayingListContext(updated as Message<boolean>, {
-      view: "everyone-selected",
-      selectedUserId,
-    });
-  }
-
   private async showEveryone(
     interaction: CommandInteraction,
     ephemeral: boolean,
@@ -1931,83 +558,6 @@ export class NowPlayingCommand {
           view: "everyone",
         });
       }
-    }
-  }
-
-  @SelectMenuComponent({ id: /^nowplaying-remove-select:\d+$/ })
-  async handleNowPlayingRemoveSelect(interaction: StringSelectMenuInteraction): Promise<void> {
-    const isEphemeral = interaction.message.flags?.has(MessageFlags.Ephemeral) ?? false;
-    const segs = assertCustomIdSegments(interaction, 1);
-    if (!segs) return;
-    const [ownerId] = segs;
-    if (await replyIfNotOwner(interaction, ownerId, "This remove prompt isn't for you.")) return;
-    const gameId = Number(interaction.values?.[0]);
-    if (!isPositiveInt(gameId)) {
-      const container = buildTextContainer("Invalid game selection.");
-      await safeReply(interaction, {
-        components: [container],
-        flags: buildComponentsV2Flags(true),
-      });
-      return;
-    }
-    const loadingContainer = buildTextContainer("Updating your Now Playing remove list...");
-    await safeUpdate(interaction, {
-      components: [loadingContainer],
-      flags: buildComponentsV2Flags(isEphemeral),
-    });
-
-    try {
-      const removed = await Member.removeNowPlaying(ownerId, gameId);
-      if (!removed) {
-        const container = buildTextContainer("Failed to remove that game (it may have been removed already).");
-        safeIgnore(safeReply(interaction, {
-          components: [container],
-          flags: buildComponentsV2Flags(isEphemeral),
-        }));
-        return;
-      }
-      safeIgnore(refreshNowPlayingListFromContext(interaction, ownerId));
-      const entries = getDisplayNowPlayingEntries(await Member.getNowPlaying(ownerId));
-      if (!entries.length) {
-        const container = buildTextContainer("Your Now Playing list is empty.");
-        const pmComponents = await withPmNowPlayingList(
-          ownerId,
-          interaction.guildId,
-          [container],
-        );
-        safeIgnore(safeReply(interaction, {
-          components: pmComponents,
-          flags: buildComponentsV2Flags(isEphemeral),
-        }));
-        return;
-      }
-      const includeImages = interaction.guildId != null;
-      const { files, thumbnailsByGameId } = await buildNowPlayingAttachments(
-        entries,
-        NOW_PLAYING_GALLERY_MAX,
-        includeImages,
-      );
-      const components = buildNowPlayingRemoveComponents(
-        entries,
-        ownerId,
-        thumbnailsByGameId,
-      );
-      const pmComponents = await withPmNowPlayingList(
-        ownerId,
-        interaction.guildId,
-        components,
-      );
-      safeIgnore(safeReply(interaction, {
-        ...buildComponentPayload(pmComponents as any, files),
-        flags: buildComponentsV2Flags(isEphemeral),
-      }));
-    } catch (err: any) {
-      const msg = extractErrorMessage(err);
-      const container = buildTextContainer(`Could not remove from Now Playing: ${msg}`);
-      safeIgnore(safeReply(interaction, {
-        components: [container],
-        flags: buildComponentsV2Flags(isEphemeral),
-      }));
     }
   }
